Fix binary plist parsing on Windows. (#26)
This fixes reading the binary plist on Windows by forcing the
binary (raw) file mode. Prior to this fix, binary plists were
either partially parsed or parsing failed completely.

This is due to the fact that reading text files on Windows is
subject to automatic line-end conversion. Therefore this bug
was never visible on MacOS/Linux/Unix.